Workspace Additions
Workspace used to move Right when holding right arrow key, which gave the illusion of a camera going left when holding right, which is a bit confusing for the player. It was like that for every direction (Up-Down; Right-Left).

Chips can now be placed out of bounds, allowing for bigger chips being placed. I just removed the collision detection.